+++ calcipy/check_for_stale_packages/__init__.py
@@ -1,1 +1,4 @@
-from ._check_for_stale_packages import check_for_stale_packages
+try:
+    from ._check_for_stale_packages import check_for_stale_packages
+except ImportError as exc:
+    raise RuntimeError("The 'calcipy[stale]' extras are missing") from exc