+++ calcipy/check_for_stale_packages/_check_for_stale_packages.py
@@ -14,8 +14,8 @@ from pydantic import BaseModel, Field, validator
 from pyrate_limiter import Duration, Limiter, RequestRate
 from shoal.can_skip import can_skip
 
+from .._log import logger
 from ..file_helpers import LOCK
-from ..log import logger
 
 try:
     import tomllib