+++ calcipy/code_tag_collector/__init__.py
@@ -1,1 +1,4 @@
-from ._collector import CODE_TAG_RE, COMMON_CODE_TAGS, write_code_tag_file
+try:
+    from ._collector import write_code_tag_file
+except ImportError as exc:
+    raise RuntimeError("The 'calcipy[tags]' extras are missing") from exc