+++ calcipy/file_helpers.py
@@ -12,7 +12,7 @@ from pathlib import Path
 from beartype import beartype
 from beartype.typing import Any, Dict, List, Optional
 
-from .log import logger
+from ._log import logger
 
 try:
     import tomllib