+++ docs/gen_ref_nav.py
@@ -34,8 +34,7 @@ for path in sorted(Path('calcipy').rglob('*.py')):
 
     mkdocs_gen_files.set_edit_path(full_doc_path, path)
 
-# add pages manually:
-# > nav["package", "module"] = "path/to/file.md"
+# FYI, to add pages manually, use: `nav["package", "module"] = "path/to/file.md"`
 
 with mkdocs_gen_files.open('reference/SUMMARY.md', 'w') as nav_file:
     nav_file.writelines(nav.build_literate_nav())