+++ tests/tasks/test_test.py
@@ -2,7 +2,7 @@ from unittest.mock import call
 
 import pytest
 
-from calcipy.tasks.invoke_helpers import use_pty
+from calcipy.invoke_helpers import use_pty
 from calcipy.tasks.test import pytest as task_pytest
 from calcipy.tasks.test import step, watch, write_json