+++ docs/docs/CHANGELOG.md
@@ -39,6 +39,7 @@
 
 ### Refactor
 
+- resolve pylint issues
 - reorder main tasks
 - resolve flake8 warnings and better manage filenames
 - apply autopep8 to fix whitespace issues